refactor: errorAttributes() 를 ErrorConfig.class 로 분리
