+++ src/main/java/techcourse/myblog/MyblogApplication.java
@@ -16,21 +16,4 @@ public class MyblogApplication {
     public static void main(String[] args) {
         SpringApplication.run(MyblogApplication.class, args);
     }
-
-    @Bean
-    public ErrorAttributes errorAttributes() {
-        return new DefaultErrorAttributes() {
-
-            @Override
-            public Map<String, Object> getErrorAttributes(WebRequest webRequest, boolean includeStackTrace) {
-                Map<String, Object> errorAttributes = super.getErrorAttributes(webRequest, includeStackTrace);
-                Throwable error = getError(webRequest);
-                if (error instanceof ValidUserException) {
-                    errorAttributes.put("errors", ((ValidUserException) error).getErrors());
-                }
-                return errorAttributes;
-            }
-        };
-    }
-
 }
