+++ src/main/java/techcourse/myblog/config/ErrorConfig.java
@@ -1,0 +1,29 @@
+package techcourse.myblog.config;
+
+import org.springframework.boot.web.servlet.error.DefaultErrorAttributes;
+import org.springframework.boot.web.servlet.error.ErrorAttributes;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.context.request.WebRequest;
+import techcourse.myblog.users.ValidUserException;
+
+import java.util.Map;
+
+@Configuration
+public class ErrorConfig {
+    @Bean
+    public ErrorAttributes errorAttributes() {
+        return new DefaultErrorAttributes() {
+
+            @Override
+            public Map<String, Object> getErrorAttributes(WebRequest webRequest, boolean includeStackTrace) {
+                Map<String, Object> errorAttributes = super.getErrorAttributes(webRequest, includeStackTrace);
+                Throwable error = getError(webRequest);
+                if (error instanceof ValidUserException) {
+                    errorAttributes.put("errors", ((ValidUserException) error).getErrors());
+                }
+                return errorAttributes;
+            }
+        };
+    }
+}
